\name{cancergenecensus}
\alias{cancergenecensus}
\docType{data}
\title{
The Sanger Cancer Gene Census
}
\description{
A data frame containing the original spreadsheet.
}
\usage{data(cancergenecensus)}
\source{
http://www.sanger.ac.uk/genetics/CGP/Census/Table_1_full_2012-03-15.xls
}
\references{
Nature Reviews Cancer 4, 177-183 (March 2004) | doi:10.1038/nrc1299
}
\examples{
data(cancergenecensus)

#cancergenecensus <- read.xls("Table_1_full_2012-03-15.xls")
}
\keyword{datasets}
